requirements.txt
attrs==19.3.0
Click==7.0
Flask==1.1.1
Flask-BasicAuth==0.2.0
Flask-HTTPAuth==3.3.0
Flask-json-schema==0.0.5
Flask-Login==0.5.0
flask-paginate==0.5.5
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.7.3
jsl==0.2.4
jsonschema==2.4.0
MarkupSafe==1.1.1
mongoengine==0.19.1
PyJWT==1.7.1
pymongo==3.10.1
pyrsistent==0.15.7
PyYAML==3.11
six==1.14.0
Werkzeug==1.0.0
zipp==2.2.0
